Refactor post request for add item
